waterfall

ECharts.waterfall — Function

waterfall(x, y)

Creates an EChart representing the increase/decrease between two consecutive points.

Methods

waterfall(x::AbstractVector, y::AbstractVector{<:Real})

Arguments

legend::Bool : display legend?

scale::Bool = false : show full Y-axis or truncated

kwargs : varargs to set any field of resulting EChart struct

Notes

Reasonable defaults set for different methods of area, such as displaying a legend when two or more series present.

using ECharts
waterfall(["Q1","Q2","Q3","Q4","Q5"], [2900, -1200, -300, -200, -900])