heatmap

ECharts.heatmap — Function

heatmap(data)

Creates an EChart heatmap.

Methods

heatmap(h::StatsBase.Histogram)

Arguments

legend::Bool = false : display legend?

show::Bool = true : show slider denoting color range

calculable::Bool = true : allow user to slide endpoints to change data displayed

kwargs : varargs to set any field of resulting EChart struct

Notes

Reasonable defaults set for different methods of area, such as displaying a legend when two or more series present.

using ECharts, StatsBase
h = fit(Histogram, (randn(10000), randn(10000)), nbins = 20)
heatmap(h)