polar

ECharts.polar — Function

polar(angle, radius)

Creates an EChart where region between points connected by a line, with polar coordinates.

Methods

polar(angle::AbstractVector{<:Union{Missing, Real}}, radius::AbstractVector{<:Union{Missing, Real}})
polar(angle::AbstractVector{<:Union{Missing, Real}},radius::AbstractArray{<:Union{Missing, Real},2})

Arguments

splitNumber::Int = 12 : number of splits shown in axis

showSymbol::Bool = false : show the dots as well as the line?

legend::Bool : display legend?

kwargs : varargs to set any field of resulting EChart struct

Notes

Reasonable defaults set for different methods of polar, such as displaying a legend when two or more series present.

polar(angle, radius)

Creates an EChart polar line chart from a 2-D array radius, where each column is a separate series. Legend is displayed by default when multiple series are present. See the primary polar method for full argument documentation.

using ECharts
t = 0:360
angle = [a / 180 * pi for a in t]
radius = [(sin(2a) * cos(2a)) for a in angle]
polar(angle, radius)